pyproject.toml
[tool.poetry]
name = "envit"
version = "0.1.0"
description = "Managing multiple environments credentials easily"
authors = ["Hassan Saadat <saadathassan34@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
cryptography = "^42.0.5"
python-decouple = "^3.8"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-cov = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ev = 'envit.cli:main'
